updated FabMaMiCo with new tasks and updated tasks

+++ FabMaMiCo.py
@@ -5,16 +5,20 @@
 #
 # This file contains FabSim definitions specific to FabDummy.
 
+import os
+
 try:
     from fabsim.base.fab import *
 except ImportError:
     from base.fab import *
 
-from fabsim.deploy.templates import template
+from rich.console import Console
+from rich.panel import Panel
+from rich.table import Table, box
 
-from plugins.FabMaMiCo.scripts.setup import MaMiCoSetup # ToDo: improve import (start with plugin-root)
-import pickle as pkl
-import copy
+from fabsim.deploy.templates import template
+from plugins.FabMaMiCo.scripts.settings import Settings
+from plugins.FabMaMiCo.scripts.setup import MaMiCoSetup
 
 # Add local script, blackbox and template path.
 add_local_paths("FabMaMiCo")
@@ -45,11 +49,33 @@ def mamico_install(config: str, **args):
     populate_env_templates()
     update_environment(args)
     with_config(config)
+    settings = Settings(FabMaMiCo_path, config)
+    mamico_setup = MaMiCoSetup(FabMaMiCo_path, config, settings)
+
+    # prepare MaMiCo locally
+    mamico_commit = mamico_setup.prepare_mamico_locally()
+    settings.update({ "mamico_commit": mamico_commit })
+    settings.delete("mamico_branch_tag_commit")
+
+    # prepare ls1-mardyn locally
+    need_ls1 = settings.get('need_ls1', False)
+    ls1_commit = mamico_setup.prepare_ls1_locally(need_ls1=need_ls1)
+    if len(ls1_commit) > 0:
+        settings.update({ "ls1_commit": ls1_commit })
+    settings.delete("ls1_branch_tag_commit")
+
+    # prepare open-foam locally
+    # need_openfoam = settings.get('need_openfoam', False)
+    # openfoam_commit = mamico_setup.prepare_openfoam_locally(need_openfoam=need_openfoam)
+    # if len(openfoam_commit) > 0:
+    #     settings.update({ "openfoam_commit": openfoam_commit })
+    # settings.delete("openfoam_branch_tag_commit")
+
+    checksum = settings.determine_md5()
+    update_environment({ "mamico_checksum": checksum })
+
+    # transfer files from config_files to remote machine
     execute(put_configs, config)
-    mamico_setup = MaMiCoSetup(FabMaMiCo_path, config)
-    mamico_setup.read_config()
-    mamico_setup.download_src_code()
-    mamico_setup.determine_md5()
 
     # prepare installation directory for MaMiCo
     run(f"mkdir -p {env.mamico_dir}")
@@ -64,10 +90,12 @@ def mamico_install(config: str, **args):
     mamico_setup.save_config_yml()
 
     # generate the command for compilation
-    compile_command = mamico_setup.generate_compile_command()
+    compile_command_ls1 = mamico_setup.generate_compile_command_ls1()
+    compile_command_mamico = mamico_setup.generate_compile_command_mamico()
 
     update_environment({
-        "compilation_command": compile_command,
+        "compilation_command_ls1": compile_command_ls1,
+        "compilation_command_mamico": compile_command_mamico
     })
 
     # run/submit compilation job
@@ -78,15 +106,17 @@ def mamico_install(config: str, **args):
         print("Compiling on login node:")
         # update job_dispatch to bash
         old_job_dispatch = env['job_dispatch']
+        old_job_name_template = env['job_name_template']
         update_environment({
             "job_dispatch": "bash",
-            "job_name_template": "${config}_${machine_name}_${task}"
+            "job_name_template": "fabmamico_${config}_${machine_name}_${task}"
         })
         # execute batch script with bash: `bash <config>_<machine>_1_mamico_install.sh`
         job(dict(script='compile'), args)
         # reset job_dispatch to its original value (for possible subsequent jobs)
         update_environment({
-            "job_dispatch": old_job_dispatch
+            "job_dispatch": old_job_dispatch,
+            "job_name_template": old_job_name_template
         })
     else:
         ###########################
@@ -112,12 +142,6 @@ def mamico_run(config: str, **args):
 
     # make sure MaMiCo is installed
     mamico_install(config, **args)
-    # TODO: if compilation as job, wait for job to finish
-
-    mamico_setup = MaMiCoSetup(FabMaMiCo_path, config)
-    mamico_setup.read_config()
-
-    print(env.cores)
 
     # submit the job
     job(dict(script='run'), args)
@@ -133,37 +157,58 @@ def mamico_run_ensemble(config: str, **args):
     """
     populate_env_templates()
     update_environment(args)
-    with_config(config)
-    execute(put_configs, config)
 
     # make sure MaMiCo is installed
     mamico_install(config, **args)
-    # TODO: if compilation as job, wait for job to finish
 
-    mamico_setup = MaMiCoSetup(FabMaMiCo_path, config)
-    mamico_setup.read_config()
+    path_to_config = find_config_file_path(config)
+    print(f"local config file path at: {path_to_config}")
+    sweep_dir = os.path.join(path_to_config, "SWEEP")
+    env.script = 'run'
+    with_config(config)
+    run_ensemble(config, sweep_dir, **args)
+
 
 @task
 @load_plugin_env_vars("FabMaMiCo")
-def mamico_scaling(config:str, **args):
+def mamico_stat(**args):
+    """
+    Show the queue of FabMaMiCo-jobs on the remote machine.
+    Calls `squeue` and filters the output for FabMaMiCo-jobs.
+    """
     populate_env_templates()
     update_environment(args)
-    with_config(config)
-    execute(put_configs, config)
+    output = run(f"squeue --format='%.10i %.9P %.62j %.10u %.8T %.13M %.9l %.6D %R' --me", capture=True)
+    output = output.split("\n")
+    output = output[0] + "\n" + "\n".join([line for line in output[1:] if "fabmamico_" in line])
+    print(output)
 
-    mamico_install(config, **args)
 
-    mamico_setup = MaMiCoSetup(FabMaMiCo_path, config)
-    mamico_setup.read_config()
+@task
+@load_plugin_env_vars("FabMaMiCo")
+def mamico_jobs_overview(**args):
+    """
+    Show the queue of FabMaMiCo-jobs on the remote machine.
+    Calls `squeue` and filters the output for FabMaMiCo-jobs.
+    """
+    populate_env_templates()
+    update_environment(args)
+    output = run("squeue --me | awk '{print $1}'", capture=True)
+    output = len(output.split("\n")[1:-1])
+    print(f"Currently having {output} jobs in the queue.")
 
-    # copy the checkpoint files to the build folder
-    run(f"cp {env.mamico_dir}/{env.mamico_checksum}/examples/CheckpointSimpleMD_10000_periodic_0.checkpoint {env.mamico_dir}/{env.mamico_checksum}/build/CheckpointSimpleMD_10000_periodic_0.checkpoint")
-    run(f"cp {env.mamico_dir}/{env.mamico_checksum}/examples/CheckpointSimpleMD_10000_reflecting_0.checkpoint {env.mamico_dir}/{env.mamico_checksum}/build/CheckpointSimpleMD_10000_reflecting_0.checkpoint")
 
-    for i in [1, 2, 4, 8, 16, 32, 64, 128]:
-        update_environment({"cores": i})
-        # EDIT THE CONFIG FILE
-        run(f"cp {env.job_config_path}/couette.xml {env.mamico_dir}/{env.mamico_checksum}/build/couette.xml")
+@task
+@load_plugin_env_vars("FabMaMiCo")
+def mamico_jobs_cancel_all(**args):
+    """
+    Cancel all `fabmamico_`-jobs on the remote machine.
+    """
+    populate_env_templates()
+    update_environment(args)
+    output = run("squeue --me | awk '{print $1}' | tail -n+2 | xargs -n 1 scancel", capture=True)
+    print(output)
+    print("All MaMiCo jobs were canceled.")
 
 
 @task
@@ -182,15 +227,37 @@ def mamico_list_installations(**args):
             expand=False,
         )
     )
+    # list all directories in the MaMiCo directory
     installations = run(f"ls {env.mamico_dir}", capture=True)
-    rich_print(
-        Panel(
-            "\n".join(installations.split()),
-            title=f"[green]Found {len(installations.split())} installations on {env.host}[/green]",
-            border_style="green",
-            expand=False
-        )
+
+    # get compilation info from each installation
+    verbose_info = []
+    if args.get("verbose", False):
+        for installation in installations.split():
+            a = run(f"cat {env.mamico_dir}/{installation}/build/compilation_info.yml", capture=True)
+            verbose_info.append(a)
+
+    # Print the installations as table
+    num_installations = len(installations.split())
+    title = f"\n[green]Found {num_installations} installation{ 's' if num_installations != 1 else '' } on {env.host}\n"\
+        f"at[/green] [white]{env.mamico_dir}[/white]"
+    table = Table(
+        title=title,
+        show_header=True,
+        show_lines=args.get("verbose", False),
+        box=box.ROUNDED,
+        header_style="blue",
     )
+    table.add_column("MD5 Checksum", style="white")
+    if args.get("verbose", False):
+        table.add_column("Compilation Settings", style="white")
+    for installation in installations.split():
+        if args.get("verbose", False):
+            table.add_row(installation, "\n".join(verbose_info.pop(0).split("\n")[:-1]))
+        else:
+            table.add_row(installation)
+    console = Console()
+    console.print(table)
 
 
 @task
@@ -233,12 +300,23 @@ def mamico_remove_all_installations(**args):
     update_environment(args)
     rich_print(
         Panel(
-            f"Please wait until all MaMiCo installations on {env.host} are removed.",
+            f"Please confirm the removal of all MaMiCo installations on {env.host} by entering 'y'.",
             title="[pink1]Cleaning MaMiCo installations[/pink1]",
             border_style="pink1",
             expand=False,
         )
     )
+    confirmation = input("Y[es] or N[o]: ").lower()
+    if confirmation != "y":
+        rich_print(
+            Panel(
+                f"Aborted the removal of MaMiCo installations on {env.host}.",
+                title="[red]Aborted[/red]",
+                border_style="red",
+                expand=False
+            )
+        )
+        return
     output = run(f"rm -rf {env.mamico_dir}/*", capture=True)
     print(output)
     rich_print(
